requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
langgraph>=0.0.15
langchain_openai>=0.1.0
langchain_core>=0.1.30
streamlit>=1.25.0
scikit-learn>=1.3.0
elasticsearch==7.17.9
pandas>=2.0.0
numpy>=1.25.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
networkx>=3.2.1
apscheduler>=3.10.4
streamlit-autorefresh==1.0.1
ollama>=0.1.7
